requirements.txt
google-generativeai>=0.2.0
streamlit>=1.0.0
requests>=2.0.0
Pillow>=9.0.0
zlib